Adjust validation of trimming loop ending points
This is a slight variation of the Oct 12 2009 adjustment to this routine
to use relative uv size to perform the validation.  Unlike the original
mod this will use the openNURBS approach if we can't get a valid range,
but the principle is the same.

The original discussion thread doesn't appear to be online any longer,
but I believe this was originally proposed upstream as a patch and they
didn't accept it.  The general idea is to check the validity using a
tolerance in the parametric space that is scaled to the overall size of
that parametric space, rather than using a ratio of the point values.